update the join background
